+++ packages/plugin-ext/src/hosted/node/plugin-reader.ts
@@ -24,6 +24,7 @@ import { BackendApplicationContribution } from '@theia/core/lib/node/backend-app
 import { PluginMetadata, getPluginId, MetadataProcessor, PluginPackage, PluginContribution } from '../../common/plugin-protocol';
 import { MetadataScanner } from './metadata-scanner';
 import { loadManifest } from '@theia/plugin-utils/lib/node/plugin-manifest';
+import { PLUGINS_BASE_PATH } from '@theia/plugin-utils/lib/constants';
 
 @injectable()
 export class HostedPluginReader implements BackendApplicationContribution {
@@ -43,7 +44,7 @@ export class HostedPluginReader implements BackendApplicationContribution {
     protected pluginsIdsFiles: Map<string, string> = new Map();
 
     configure(app: express.Application): void {
-        app.get('/hostedPlugin/:pluginId/:path(*)', async (req, res) => {
+        app.get(`/${PLUGINS_BASE_PATH}/:pluginId/:path(*)`, async (req, res) => {
             const pluginId = req.params.pluginId;
             const filePath = req.params.path;