+++ packages/plugin-utils/src/node/read-grammars.spec.ts
@@ -18,7 +18,7 @@ import { expect } from 'chai';
 import * as fs from 'fs-extra';
 import * as os from 'os';
 import * as path from 'path';
-import { readGrammarFromDisk, getGrammarContributionValidationError } from '../node/read-grammars';
+import { readGrammarFromDisk, getGrammarContributionValidationError } from './read-grammars';
 
 describe('readGrammarFromDisk', () => {