+++ packages/plugin-utils/src/plugin-manifest.spec.ts
@@ -15,10 +15,10 @@
 // *****************************************************************************
 
 import { expect } from 'chai';
-import { UNPUBLISHED, VSCODE_BUILTIN_NAME_PREFIX } from '../constants';
-import { prepareLoadedManifest, stripVscodeBuiltinNamePrefix } from '../plugin-manifest';
-import { loadManifest } from '../node/plugin-manifest';
-import type { PluginManifest } from '../manifest-types';
+import { UNPUBLISHED, VSCODE_BUILTIN_NAME_PREFIX } from './constants';
+import { prepareLoadedManifest, stripVscodeBuiltinNamePrefix } from './plugin-manifest';
+import { loadManifest } from './node/plugin-manifest';
+import type { PluginManifest } from './manifest-types';
 import * as fs from 'fs-extra';
 import * as os from 'os';
 import * as path from 'path';